+++ src/components/Base/index.ts
@@ -1,0 +1,1 @@
+export { Base } from './Base'
