+++ src/components/Canvas/Canvas.tsx
@@ -1,28 +1,19 @@
-import { NearestFilter, TextureLoader } from 'three'
-import { Canvas as R3Canvas, useLoader } from '@react-three/fiber'
-import { Edges, OrbitControls } from '@react-three/drei'
+import { Canvas as R3Canvas } from '@react-three/fiber'
+import { OrbitControls } from '@react-three/drei'
+
+import { Base } from '../Base'
+import { Cargo } from '../Cargo'
 
 import './Canvas.css'
 
 export function Canvas() {
-  const notebookTexture = useLoader(TextureLoader, '/squares.png')
-  notebookTexture.minFilter = NearestFilter
-  notebookTexture.magFilter = NearestFilter
-
   return (
     <section className="canvas">
       <R3Canvas>
-        <mesh rotation={[-Math.PI / 2, 0, 0]}>
-          <planeGeometry args={[96, 54]} />
-          <meshStandardMaterial map={notebookTexture} />
-        </mesh>
         <ambientLight intensity={1} />
         <directionalLight position={[5, 5, 5]} intensity={1} />
-        <mesh position={[0, 0.5, 0]}>
-          <boxGeometry />
-          <meshStandardMaterial color={0xffa500} />
-          <Edges linewidth={2} threshold={1} color={0xffffff} />
-        </mesh>
+        <Base />
+        <Cargo />
         <OrbitControls
           minDistance={10}
           maxDistance={50}
