+++ src/components/Cargo/index.ts
@@ -1,0 +1,1 @@
+export { Cargo } from './Cargo'
